requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.7
django-braces==1.11.0
django-crispy-forms==1.7.0
django-debug-panel==0.8.3
django-debug-toolbar==1.9.1
django-pyodbc-azure==1.11.0.0
django-tables2==1.14.2
idna==2.6
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
pip==9.0.1
PyJWT==1.5.3
pyodbc==4.0.21
python3-openid==3.1.0
pytz==2017.3
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==0.8.0
setuptools==36.7.2
six==1.11.0
social-auth-app-django==2.0.0
social-auth-core==1.5.0
sqlparse==0.2.4
urllib3==1.22
wfastcgi==3.0.0
